Fix image updating issue

+++ local-addon/pm_accounting/models/pm_accounting.py
@@ -69,7 +69,7 @@ class PaymentIcon(models.Model):
 
     def write(self, vals):
       # Temporarily fixing image issue when update a record
-      if vals['image']:
+      if 'image' in vals and vals['image']:
           self.env.cr.execute("""DELETE FROM ir_attachment WHERE res_model = '%s' AND res_field = '%s' AND res_id = %d""" % (self._name, 'image', self.id))
 
       return super(PaymentIcon, self).write(vals)
