+++ local-addon/pm_culinary/models/pm_brand.py
@@ -14,7 +14,7 @@ class PmBrand(models.Model):
 
     def write(self, vals):
       # Temporarily fixing image issue when update a record
-      if vals['logo']:
+      if 'logo' in vals and vals['logo']:
           self.env.cr.execute("""DELETE FROM ir_attachment WHERE res_model = '%s' AND res_field = '%s' AND res_id = %d""" % (self._name, 'logo', self.id))
 
       return super(PmBrand, self).write(vals)
