+++ local-addon/pm_leads/models/lead.py
@@ -614,7 +614,7 @@ class Lead(models.Model):
 
     def write(self, vals):
       # Temporarily fixing image issue when update a record
-      if vals['image']:
+      if 'image' in vals and vals['image']:
           self.env.cr.execute("""DELETE FROM ir_attachment WHERE res_model = '%s' AND res_field = '%s' AND res_id = %d""" % (self._name, 'image', self.id))
 
       return super(Lead, self).write(vals)
